WIP range-diff: support reading mbox files
TODO:
	- document
	- split out the `mbox:-` code
	- add a test (probably by using `git format-patch --stdout`)

Internally, the `git range-diff` command spawns a `git log` process and
parses its output for the given commit ranges.

This works well when the patches that need to be compared are present in
the local repository in the form of commits.

In scenarios where that is not the case, the `range-diff` command is
unhelpful.

For example, if one desired to compare commits in a local repository to
patches that were sent to the Git mailing list, it would be necessary to
apply the patches locally first, which requires finding an appropriate
base commit or adjusting the patches manually so that they apply. This
is a lot of effort given that `range-diff` would then go and internally
re-generate the patches, anyway.

Let's offer a way to read those patches from pre-prepared MBox files
instead when an argument "mbox:<filename>" is passed instead of a commit
range.

This addresses https://github.com/gitgitgadget/git/issues/207

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>