Introduce `core.useBusyBox` to use BusyBox' ash as shell
BusyBox' ash is a bit tricky because it needs to be called as `busybox
sh` (although there is sometimes a hard-linked `ash` in a location like
`<prefix>/libexec/busybox/ash`, but that is not guaranteed), i.e. there
is not necessarily single absolute path to use as `SHELL_PATH`.

Therefore, it is not an option to ask users who wish to use BusyBox'
`ash` as their shell to "just set SHELL_PATH".

Let's offer `core.useBusyBox` for that purpose: if set to `true`, Git
will use BusyBox' built-in `ash` whenever it needs to call a Unix shell
(e.g. for commands still implemented as shell script, and hooks and
aliases). .

To use it when running Git's tests, users can define the environment
variable `GIT_TEST_USE_BUSYBOX` to a non-empty string.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>