clean_on_exit: avoid memory corruption
When we're asking the `run_command()` machinery to clean up child
processes, the `struct child_process` instance in question is added to a
list. We require instances that are added to this list to survive until
the `atexit()` handlers are called.

However, often we just allocate that structure on the heap, e.g. when
executing aliases. This leads to problems when looking at the
`clean_on_exit_handler` attribute of that structure, as the memory might
very well have been used for different purposes in the meantime.

This instance is only needed, however, if `clean_on_exit_handler` is
non-`NULL` to begin with.

Let's not require the `struct child_process` instance at all if the
handler is `NULL`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>