merge-ort: allow rename detection to be disabled
This is most likely a really bad idea; lack of rename detection results
in suboptimal handling of merges with users being left to pick up the
pieces.  The ability to disable rename detection existed in
merge-recursive solely because its performance was abysmal.  merge-ort
is generally faster with rename detection than merge-recursive is
without rename detection, basically obviating the need for such an
option entirely.  And if merge-ort isn't fast enough with rename
detection active, I want to know about those cases to determine if there
are further optimizations available.  I could not find any testcases in
months of use by dozens of folks and in my own searching where
merge-ort's merges are not essentially instantaneous.  The only thing
that has been found to be slow, is the updating of the index and working
copy in microsoft-sized repositories, and turning off rename detection
does not help with that problem one iota.

HOWEVER, having this patch available for Stolee so he can deploy ort
early and have confidence that if anything goes awry he can just quickly
disable the rename detection probably has some merit.  It's just a
reassurance factor for him, not something I'm going to upstream.

Signed-off-by: Elijah Newren <newren@gmail.com>