fsmonitor--daemon: special-case .git's NTFS short names
It is a peculiar thing that `ReadDirectoryChangesW()` does not
canonicalize file names before reporting them. For example, when the
`.git/` directory is deleted via its short name (which is typically
`GIT~1`), _that_ name gets reported.

In order to _still_ handle this correctly, let's add a work-around
specifically for this use case.

Note: it is rare these days that files are accessed via their short
names. So let's not worry too much about other instances of this
problem.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>