artifacts-tar: include Perl artifacts unless `NO_PERL` is defined
We forgot to include the Perl artifacts! This was not noticed earlier
because currently our only use for this Makefile target is the
Windows-specific part of our CI builds, where we explicitly omitted all
Perl artifacts.

But we are about to allow optionally testing all of the Perl stuff, too,
and hence we need `artifacts-tar` to respect the absence of the
`NO_PERL` setting.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>