ci(windows): do test the Perl stuff when targeting the `main` branch
We really should make sure that the Perl stuff works. It is not _so_
important as to slow down each and every CI build, but when targeting
the `main` branch we should invest the required time and effort.

For good measure, we include the `shears/master` branch because it
provides a glimpse of what is about to trickle down from upstream into
our `main` branch. Testing the Perl parts here, too, gives us a chance
to react early enough.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>