dir: add path_matches_cone_mode_pattern_list()
When matching against a generic pattern list, the 'basename' is
important for some patterns. However, it and the 'dtype' parameter are
irrelevant for cone mode sparse-checkout patterns. If we know that we
are working with cone mode patterns from the start, then we can speed up
the pattern check slightly by not computing the 'basename'.

In many existing consumers, the 'basename' is already known from
context, but some new consumers we compute this on-demand. A future
change will add more calls that do not have the 'basename' from context
and would need to compute it for many cache entries in a tight loop.
Avoid this problem by creating the new
path_matches_cone_mode_pattern_list() method.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>