update-index: integrate with sparse index
Enable usage of the sparse index with `update-index`. Most variations of
`update-index` work without explicitly expanding the index (due to the
implicit expansion performed when attempting to find an existing entry).
However, in order to prevent manually creating a sparse directory cache
entry (current behavior is to throw an error when presented with a directory
using `--cacheinfo`), the index is expanded before the the entry is added.


Signed-off-by: Victoria Dye <vdye@github.com>