read-tree: narrow scope of index expansion for `--prefix`
When `unpack_trees` is provided with a prefix, the tree traversal that occurs
in it requires that the base directory path given to it is a traversable
directory in the index. In order to adhere to this requirement, the index is
expanded if the given prefix is equal to or a subdirectory of a sparse
directory in the index.

Signed-off-by: Victoria Dye <vdye@github.com>