blame: enable and test the sparse index
Enable the sparse index for the 'git blame' command. The index was already not
expanded with this command, so the most interesting thing to do is to add tests
that verify that 'git blame' behaves correctly when the sparse index is enabled
and that its performance improves. More specifically, these cases are:

1. The index is not expanded for `blame` when given paths in the sparse
checkout cone at multiple levels.

2. Performance measurably improves for
`blame` with sparse index when given paths in the sparse checkout cone at
multiple levels.

The following are the performance results comparing the branch with sparse
index enabled for `blame` with the vfs-2.33.0 branch in the microsoft/git
repository:

```
Test                                       msft/vfs-2.33.0   sparse-index-blame
-----------------------------------------------------------------------------------
2000.62: git blame f2/f4/a (full-v3)        0.31(0.21+0.09)   0.32(0.21+0.10) +3.2%
2000.63: git blame f2/f4/a (full-v4)        0.29(0.19+0.08)   0.31(0.21+0.08) +6.9%
2000.64: git blame f2/f4/a (sparse-v3)      0.55(0.38+0.14)   0.23(0.14+0.07) -58.2%
2000.65: git blame f2/f4/a (sparse-v4)      0.57(0.40+0.16)   0.23(0.15+0.07) -59.6%
2000.66: git blame f2/f4/f3/a (full-v3)     0.77(0.56+0.20)   0.85(0.60+0.22) +10.4%
2000.67: git blame f2/f4/f3/a (full-v4)     0.78(0.56+0.20)   0.81(0.59+0.21) +3.8%
2000.68: git blame f2/f4/f3/a (sparse-v3)   1.07(0.78+0.28)   0.72(0.52+0.19) -32.7%
2000.99: git blame f2/f4/f3/a (sparse-v4)   1.05(0.78+0.26)   0.73(0.51+0.19) -30.5%
```

We do not include paths outside the sparse checkout cone because blame
currently does not support blaming files outside of the sparse definition.
Attempting to do so fails with the following error:

fatal: no such path '<path outside sparse definition>' in HEAD

Signed-off-by: Lessley Dennington <lessleydennington@gmail.com>