WIP: fsck: teach --connectivity-only to play well with --name-objects
Oddly enough, the `parse_tree_gently()` function _dies_ on corrupt
objects (despite its name suggesting otherwise).

This is a problem when running `git fsck --name-objects` because Git is
only showing the SHA before it dies, and does not name the object in an
intuitive, human-readable way at all.

Let's fix this.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>