WIP: allow fine-grained invalidation of FSCache entries
This does not work quite yet: when creating or deleting an intermediate
directory (`dir` in `dir/subdir/file`), the deeper levels are not yet
handled correctly.

It _might_ be possible to do this in an even finer-grained manner, by
specifically calling `mingw_lstat()` and copying the information (it's a
bit more complicated for entities that did not exist before, or that no
longer exist).

Also: the parts that are about the lstat cache need to be split out of
this patch.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>