ci: retire the concept of the `cache_dir`
Back in the Travis CI days, we looked for a way to skip re-runs of
already-successful builds for tree-same commits, to save on time and
energy (we have to start _somewhere_ to save our home planet).

The strategy we settled on was to store a file in a "cache directory",
i.e. a directory that was preserved between CI runs.

When we moved to Azure Pipelines, we did not really find any good way to
preserve that strategy (for a while, we tried to mount an Azure file
share, using a repository secret, but we abandoned that idea when moving
to GitHub Actions).

On GitHub Actions, we now use the strategy of enumerating the latest
successful runs and looking for the current commit/tree to figure out
whether a re-run is unnecessary (see 876511ef614 (Merge branch
'js/ci-ghwf-dedup-tests', 2020-10-30) for details).

In other words, we haven't used the cache directory approach ever since
abandoning our Travis CI definition as per 4a6e4b96026 (CI: remove
Travis CI support, 2021-11-23).

Even if we shold desire to reinstate Travis CI support, we are more
likely to imitate the current approach of using API calls to figure out
recent successful runs than to reinstate the cache directory strategy
(which does have a racy problem with concurrently-run builds, after
all).

So let's remove the last remnants of the cache directory code from our
CI scripts.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>