t2501: guard two test cases that cannot work in pure Windows environments
On Windows, a directory simply cannot be removed if there is even one
process whose current directory it is.

Two test cases specifically test for this, let's skip them when running
with BusyBox-w32's `ash.exe`.

Side note: This operation succeeds in Cygwin due to a nasty hack that
allows Linux-like behavior (at the expense of messing with Windows'
internals and all the consequences that come with that).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>