subtree: respect TEST_SHELL_PATH, too
In 3f824e91c848 (t/Makefile: introduce TEST_SHELL_PATH, 2017-12-08), we
taught the regular Git test suite a new trick: to use a different shell
to build Git than to run Git's test suite.

All of `contrib/` was left out, though, including `subtree`. As we want
to run the `subtree` tests in Git for Windows, and as BusyBox needs to
make use of that `TEST_SHELL_PATH` trick, we need to teach that trick at
least to that subdirectory in `contrib/`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>