t2205: avoid `local` outside of a function
In 27128996b88f (dir: traverse into repository, 2022-06-16), four test
cases were added that use the keyword `local` to declare a variable
inside a subshell.

But this code is not run within a function, therefore BusyBox' `ash`
protests loudly. Let's just drop the `local` attributes, they are not
necessary anyway because we are inside a subshell.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>