fixup! scalar: verify that we can use a GVFS-enabled repository
As of 65f6a9eb0b9b (scalar: constrain enlistment search, 2022-08-18),
`scalar` no longer detects enlistments when run from subdirectories that
are _parallel_ to the repository.

Account for that in the test case that specifically tried to verify that
we detect enlistments even when run outside the repository and not at
the enlistment root itself.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>