FOR-TESTING: reintroduce Nano Server breakage
To verify that the `nano-server.yml` GitHub workflow would catch a
regression, let's intentionally re-introduce the bug that we fixed in
v2.39.0.

This reverts the mimalloc commits 948a0c4 (Lazy-load `psapi.dll`,
2019-10-20), edb181c377b (Lazy-link to `bcrypt`, 2022-08-12) and
2c40e0d02fc (Allow running in Windows Nano Server containers,
2022-10-18), and adds `-lpsapi` and `-lbcrypt` to the linker flags,
which in turn requires a full SDK (because the minimal subset of the SDK
does not include `libpsapi.a`).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>