Start re-implementing `subtree` as a built-in
It is currently a shell script. Such scripts frequently pose problems on
Windows, from poor performance characteristics to falsely quarantined
"malware" to problems with the POSIX emulation layer provided by the
MSYS2 runtime.

Therefore it is a good idea to avoid shell scripts in Git's production
code altogether.

"But why `subtree`?" you may ask. "Isn't it in `contrib/`?" you may ask.
Why yes, it is. But for historical reasons, it is included in Git for
Windows, and did provide for a semi-steady trickle of bug reports.

So let's re-implement it in C.

This commit only provides a mere shell (pun intended) of a built-in. The
functionality will be implemented later.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>