test-tool: offer to generate insanely-large synthetic pack files
To study Git at scale, and to keep it working even for the largest
hosted repositories, we need a way to generate pack files containing
_many_ objects.

This new helper generates such synthetic pack files, fast.

To do so, it side-steps most of Git's regular machinery and even avoids
the time tax of deflating (and later on inflating) the objects.

Instead, it hard-codes objects using the uncompressed format so that
objects can be generated _really_ fast.

On my laptop this tool can generate a pack with 215 million objects in
slightly over 40 seconds.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>