ci(win test): deal better with hanging tests
When tests hang, we do not see anything in the output, and the job just
times out after 6 hours (!). To add insult to injury, the timeout
prevents the `failed-tests-windows` artifact from being uploaded that is
supposed to be uploaded to help diagnose the issue better.

So let's first establish a generous timeout of 20 minutes for the
`windows-test` matrix jobs (they typically take less than 10 minutes),
and then handle the timed-out tests in addition to the failed ones.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>