ci: be a bit more tenacious gathering failed tests' stuff
When tests fail in CI, there is no easy way to find out what is going
wrong unless the logs have the information, or at least artifacts
attached to the failed CI run.

Therefore we must try a little bit harder and not stop when a file
expected to be present could not be found, but just upload whatever
relevant files we can find about that failed test.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>