ci(linux32): do make v4 of `actions/upload-artifact` work
The big problem here is that this is a Javascript-based Action, and the
way the Actions runner handles these with Docker containers is to map
the `externals/` directory that contains the runner's copy of node 20 as
host volume into the container, where it is then used to run the Action
via `/__e/node20/bin/node <args>`. But that cannot work because the
runner is amd64, and so is its copy of node 20, while the container
runs in x86 mode, and hence _must_ fail. The symptom is:

  exec /__e/node20/bin/node: no such file or directory

We just don't see it in the Git project because the only time this
Action is used is to upload an artifact in case the test suite fails,
and this is not happening all that often, and when it does, only few
people have a look.

Let's address this by overriding the amd64 version of node 20 with an
x86 one (which is "officially unofficial" by virtue of being hosted on
unofficial-builds.nodejs.org).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>