ci(linux32): do make Javascript Actions work
In 9c261856c91 (ci: use regular action versions for linux32 job,
2024-09-12), a hack was implemented to let the 32-bit container make use
of the 64-bit node 20 provided by the Actions runner.

This, however, runs the risk of using 64-bit executables when we
specifically chose a Docker image that only contains 32-bit bits and
pieces so that would not happen.

Let's address this instead by overriding the amd64 version of node 20
with an x86 one (which is "officially unofficial" by virtue of being
hosted on unofficial-builds.nodejs.org).

This partially reverts 9c261856c91 (ci: use regular action versions for
linux32 job, 2024-09-12), keeping the updates to Javascript Actions.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>