WIP: git add -p: allow stashing hunks
In the default `git add -p` mode, users can selectively stash hunks.

In common work scenarios, I often encountered situations where I would
love to stash (or simply discard) hunks instead.

This idea has been brought up on the Git mailing list before, and it is
an idea worth putting effort behind.

This patch is utterly incomplete!

In particular, the reassemble-and-apply phase needs to be accompanied by
a preceding step that would reassemble the "to-stash" hunks and stash
them.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>