amend! amend! WIP: add a test for `push --path-walk` not reusing deltas
TO-SQUASH: add a test for `push --path-walk` not reusing deltas

Let's also add a test case that demonstrates that `git -c
pack.usePathWalk=true push` avoids reusing deltas.

We do this by constructing a pack with a horrendously inefficient delta
object, then verifying that the pack on the receiving side of the `push`
fails to have such an inefficient delta.

The test case would probably be a lot more readable if hex numbers were
used instead of octal numbers, but alas, `printf "\x<hex>"` is not
portable, only `printf "\<octal>"` is. For example, dash's built-in
`printf` function simply prints `\x` verbatim.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>