fixup! Import the source code of mimalloc v2.1.2
To support mimalloc-enabled builds also on Linux and on macOS, we need
to ship the respective backends, too.

This helped me diagnose "mimalloc vs reftable" problems much quicker
because Git's birthplace *is* Linux, and its support for debugging on
that platform is still highly unmatched by its support for any other
platform.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>